refactor: move SpinalSimMultiThreadingTest
